stock_sentiment_bot/requirements.txt
# Stock Sentiment Bot Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0

# Alpaca API (trading + market data)
alpaca-py>=0.21.0

# Reddit scraping
praw>=7.7.0

# News/RSS parsing
feedparser>=6.0.0

# NLP / Sentiment
textblob>=0.17.0
vaderSentiment>=3.3.2

# Database
sqlalchemy>=2.0.0

# HTTP requests (for news APIs)
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Technical indicators
ta>=0.11.0

# Telegram notifications
python-telegram-bot>=20.0

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Async support
asyncio-throttle>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Linting (dev)
ruff>=0.1.0
mypy>=1.5.0

# Rate limiting
ratelimit>=2.2.1
